// src/pages/Admin/Mapdanda/types.ts
export interface MapdandaValues {
  groundCoverageStandard: string;
  groundCoverageActual: string;
  groundCoverageRemarks: string;
  totalFloorAreaStandard: string;
  totalFloorAreaActual: string;
  totalFloorAreaRemarks: string;
  tallaStandard: string;
  tallaActual: string;
  tallaRemarks: string;
  rowStandard: string;
  rowActual: string;
  rowRemarks: string;
  setbackStandard: string;
  setbackActual: string;
  setbackRemarks: string;
  highTesionStandard: string;
  highTesionActual: string;
  highTesionRemarks: string;
  riverStandard: string;
  riverActual: string;
  riverRemarks: string;
  farStandard: string;
  farActual: string;
  farRemarks: string;
}